requirements.txt
google-genai
feedparser